Added preliminary methods for interplay with xf_lowqual
No tests were added- out of scope for this quarter's goals, but we are getting a foot in the door by adapting these methods from main_indev to give a better image of what this will do. The final product is essentially the same and fully functional, but no conformation of tests have been added.