Spent 5 hours trying to solve the memory managment and runtime issues
Turns out, Pod5 is the one hogging all the memory. I made a quick cut just to reduce the amount of data the thing was processing that can be dealt with later, but for now, it's still obscene. May be worht looking into better methods of parsing pod5.

Updated shannon entropies to output less memory-intensive objects from the methods, which can be converted back to dataframes if nessecary.