Adjusted Aggregate_Reads.py to be able to extract features and export to either batched parquet or json
As written above, aggregate reads is now able to directly calculate the features for an item, export batches to parquet, or to json. Also added some logic to keep adding to the directory rather than wiping it clean if the process is interrupted or stopped.
Also, began working on developing ML model on subset of data.